apps/requirements.txt
langchain-google-genai
langchain==0.1.0
langgraph==0.0.17
pydantic>=2.5.2 